requirements.txt
confluent-kafka>=2.6,<3
websocket-client==1.8.0
